requirements.txt
kafka-python==2.0.2
confluent-kafka==2.3.0
avro==1.11.3
requests==2.31.0
python-dotenv==1.0.0
matplotlib==3.8.2
pandas==2.1.4
numpy==1.26.2
cassandra-driver==3.29.1

